Fixing and refactored unit tests on wp_cli and wp_tools modules.
